17851 ipadm_create_addr() returns IPADM_NOTSUP
If libipadm is 64-bit but /lib/inet/ipmgmtd is 32-bit,
ipmgmt_setaddr_handler() and i_ipadm_addr_persist_nvl() both
use ipmgmt_setaddr_arg_t which has size 0x10 for 64-bit and
0xc for 32-bit:

   mdb amd64/libipadm.so.1
   > ::sizeof struct ipmgmt_setaddr_arg_s
   sizeof (struct ipmgmt_setaddr_arg_s) = 0x10

   mdb i386/libipadm.so.1
   > ::sizeof struct ipmgmt_setaddr_arg_s
   sizeof (struct ipmgmt_setaddr_arg_s) = 0xc

Due to this ipmgmt_setaddr_handler() wrongly determines address
of nvlist header:

    nvlbuf = (char *)argp + sizeof (ipmgmt_setaddr_arg_t);

https: //illumos.org/issues/17851

+++ usr/src/lib/libipadm/common/ipadm_ipmgmt.h
@@ -220,7 +220,7 @@ typedef struct ipmgmt_initif_arg_s {
 	ipmgmt_door_cmd_type_t	ia_cmd;
 	uint32_t	ia_flags;
 	sa_family_t	ia_family;
-	size_t		ia_nvlsize;
+	uint32_t	ia_nvlsize;
 	/* packed nvl follows */
 } ipmgmt_initif_arg_t;
 
@@ -236,7 +236,7 @@ typedef struct ipmgmt_ipmp_update_arg_s {
 typedef struct ipmgmt_setaddr_arg_s {
 	ipmgmt_door_cmd_type_t	ia_cmd;
 	uint32_t		ia_flags;
-	size_t			ia_nvlsize;
+	uint32_t		ia_nvlsize;
 	/* packed nvl follows */
 } ipmgmt_setaddr_arg_t;
